move switch buttons to navbar
allow CVRs to be signed by admin, too.
