+++ libs/auth/jest.config.js
@@ -5,6 +5,14 @@ const shared = require('../../jest.config.shared');
  */
 module.exports = {
   ...shared,
+  coverageThreshold: {
+    global: {
+      statements: 100,
+      branches: 99.75,
+      functions: 100,
+      lines: 100,
+    },
+  },
   coveragePathIgnorePatterns: [
     'src/cypress.ts',
     'src/index.ts',
